+++ CHANGELOG.md
@@ -33,3 +33,8 @@ December 9th, 2023
 - Improved support for managing recordings when multiple capture devices are used simultaneously.
 - Updated permission verification check.
     - Not fatal errors have been turned into warnings that don't prevent the rest of the page from loading.
+- The following instance configuration values can now be modified from the Optic web interface.
+    - `dashcam>saving>unsaved_history_length`
+    - `dashcam>capture>opencv>stamps>main>message_1`
+    - `dashcam>capture>opencv>stamps>main>message_2`
+- Added hover text to several configuration values in the instance configuration web interface.